% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bril.R
\name{bril}
\alias{bril}
\title{Bootstrap and Refine Iterative Locator}
\usage{
bril(
  data,
  maxIterations = NULL,
  minUnassigned = 0.1,
  method = "Projection",
  alpha = 0.5,
  testUnimodal = "DIP",
  threshUnimodal = 0.05,
  distUnimodal = "Euclidean",
  testNormal = "Mardia",
  threshNormal = 0.05,
  distNormal = "MCD",
  trimmedPerFilteringIteration = 1,
  exitWhenUnimodal = FALSE,
  debug = FALSE,
  warnings = FALSE
)
}
\arguments{
\item{data}{Matrix of numerical values containing the observations (one per
row, with two columns for X and Y coordinates)}

\item{maxIterations}{Maximum number of iterations performed by the algorithm
(set to NULL or 0 for unlimited number)}

\item{minUnassigned}{Numerical value between 0 and 1 (default: 0.1),
providing the proportion of unasnigned samples from \code{data} below which the
algorithm will terminate.}

\item{method}{Method to use. Valid options are "MCD" and "MVE" for convex
body minimizers, or "L2", "Lui", "Mahalanobis", "Oja", "Projection"
(default), "Spatial" and "Tukey" for depth functions}

\item{alpha}{Proportion of samples trimmed at each iteration of the recursive
median estimate (numerical value between 0 and 1, default: 0.5), see
\code{\link[=median_rec]{median_rec()}}}

\item{testUnimodal}{Statistical test used for unimodality. Valid options are
only "DIP", see \code{\link[=filter_outliers]{filter_outliers()}}}

\item{threshUnimodal}{Threshold of significance for the unimodality test
(default: 0.05)}

\item{distUnimodal}{Distance metric used for ordering the samples in the
unimodal filtering. Valid options are "Euclidean" (default), or "MCD",
"MVE", and "OGK" for robust distances. "Euclidean" is strongly advised for
unimodality tests.}

\item{testNormal}{Statistical test used for normality. Valid options are
"Mardia" (default), "Kurtosis", "Skewness", "KS", "KS-adj", "Shapiro",
"Lillie" and "Chisq", see \code{\link[=filter_outliers]{filter_outliers()}}}

\item{threshNormal}{Threshold of significance for the normality test
(default: 0.05)}

\item{distNormal}{Distance metric used for ordering the samples in the normal
filtering. Valid options are "Euclidean", or "MCD" (default), "MVE", and
"OGK" for robust distances. Robust distances are strongly advised for
normality tests.}

\item{trimmedPerFilteringIteration}{Number of samples trimmed at each
iteration of the unimodality and normality filtering (default: 1), see
\code{\link[=filter_outliers]{filter_outliers()}}}

\item{exitWhenUnimodal}{Logical value. \code{TRUE} will terminate the execution of
the algorithm as soon as an unimodal subset is encountered on the start of
an iteration. \code{FALSE} (default) will let that last iteration proceed before
exiting}

\item{debug}{Logical value. \code{TRUE} will compute all p.values in the filtering
steps (even after the  p.values exceed the threshold, see
\code{\link[=plot.BRIL.Filtering]{plot.BRIL.Filtering()}})}

\item{warnings}{Logical value, to display the warnings and errors caught}
}
\value{
The function returns an S3 object of type \code{BRIL} containing the
following values:
\item{\code{call}}{Parameters of the call to the function (i.e. \code{data},
\code{maxIterations}, \code{minUnassigned}, \code{method}, \code{alpha}, \code{testUnimodal},
\code{threshUnimodal}, \code{distUnimodal}, \code{testNormal}, \code{threshNormal},
\code{distNormal}, \code{trimmedPerFilteringIteration}, and \code{exitWhenUnimodal})}
\item{\code{iterations}}{A list of each iteration, containing the two filtering
procedures performed: \code{filteringUnimodal} and \code{filteringNormal} (both being
S3 object of class \code{BRIL.Filtering}, see \code{\link[=filter_outliers]{filter_outliers()}})}
\item{\code{nbClusters}}{Number of groups encountered (one per iteration)}
\item{\code{labels}}{Labels of the groups encountered (cooresponding to the
number of the iteration they were identified from)}
\item{\code{clustersCenters}}{Matrix containing the coordinates of the centers
of each group (row-wise) the filtering}
\item{\code{clustersSizes}}{Array with the number of sample in each group}
\item{\code{mainCluster}}{Index of the group identified as main mode}
}
\description{
Robust Estimate of Mode in Multivariate Distribuition
}
\examples{

XY <- rbind(
  mvtnorm::rmvnorm(300, c(0,0), diag(2)*3-1),
  mvtnorm::rmvnorm(100, c(15,20), diag(2)),
  mvtnorm::rmvnorm(150, c(-10,15), diag(2)*2-0.5),
  mvtnorm::rmvnorm(200, c(5,5), diag(2)*200)
)

res <- bril(XY, debug=TRUE)
print(res)

# Plot the mode and groups encountered
plot(res)

# Plot the different iterations (interactive)
\dontrun{
plot(res, "iterations")
}
}
\seealso{
\code{\link[=plot.BRIL]{plot.BRIL()}}, \code{\link[=print.BRIL]{print.BRIL()}}, \code{\link[=median_rec]{median_rec()}}, \code{\link[=median_mv]{median_mv()}}, \code{\link[=filter_outliers]{filter_outliers()}}, \code{\link[=depth_values]{depth_values()}}
}
