% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/median_mv.R
\name{median_mv}
\alias{median_mv}
\title{Multivariate Median}
\usage{
median_mv(data, method = "Projection", sampleMedian = TRUE, warnings = FALSE)
}
\arguments{
\item{data}{Matrix of numerical values containing the observations (one per
row, with two columns for X and Y coordinates)}

\item{method}{Method to use. Valid options are "CW", "MCD", "MVE", "L2",
"Lui", "Mahalanobis", "Oja", "Projection" (default), "Spatial" and "Tukey"}

\item{sampleMedian}{Logical value. If \code{TRUE} (default), the function will
return the Sample Median (observation from \code{data} with the highest depth).
If \code{FALSE}, it will return the classic Multivariate Median (point in space
with the highest depth) when applicable (methods "Oja", "Turkey" and
"Spatial").}

\item{warnings}{Logical value, to display the warnings and error caught by
the underlying functions}
}
\value{
Coordinates of the Multivariate Median
}
\description{
Computes the Multivariate Median of the distribution provided in \code{data}
(depth computations rely on the packages
\href{https://cran.r-project.org/web/packages/depth/index.html}{\code{depth}},
\href{https://cran.r-project.org/web/packages/OjaNP/index.html}{\code{OjaNP}} and
\href{https://cran.r-project.org/web/packages/ddalpha/index.html}{\code{ddalpha}},
while "MCD" and "MVE" use the package
\href{https://cran.r-project.org/web/packages/MASS/index.html}{\code{MASS}}).
}
\examples{

## Example 1

# illustrative data
XY <- rbind(
  mvtnorm::rmvnorm(300, c(0, 0), diag(2)),
  mvtnorm::rmvnorm(100, c(15, 20), diag(2) * 3 - 1),
  mvtnorm::rmvnorm(150, c(-10, 15), diag(2) * 2 - 0.5),
  mvtnorm::rmvnorm(200, c(5, 5), diag(2) * 200)
)

# compute median
m <- median_mv(XY, method = "Liu")

# plot results
plot(XY, asp = 1, xlab = "X", ylab = "Y")
graphics::points(points(m[1], m[2], col = "red", pch = 3, cex = 1.5, lwd = 3))

## Example 2

median_mv(XY, method = "L2")
median_mv(XY, method = "Tukey", sampleMedian = TRUE)
median_mv(XY, method = "Tukey", sampleMedian = FALSE)

}
\seealso{
\code{\link[=median_rec]{median_rec()}}, \code{\link[=depth_values]{depth_values()}}, \code{\link[=bril]{bril()}}
}
