% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/median_rec.R
\name{median_rec}
\alias{median_rec}
\title{Recursive estimate of central location}
\usage{
median_rec(
  data,
  method = "Projection",
  alpha = 0.5,
  maxIterations = NULL,
  warnings = FALSE
)
}
\arguments{
\item{data}{Matrix of numerical values containing the observations (one per
row, with two columns for X and Y coordinates)}

\item{method}{Method to use. Valid options are "MCD" and "MVE" for convex
body minimizers, or "L2", "Lui", "Mahalanobis", "Oja", "Projection"
(default), "Spatial" and "Tukey" for depth functions}

\item{alpha}{Proportion of samples trimmed at each iteration (numerical value
between 0 and 1, default: 0.5)}

\item{maxIterations}{Set to a positive integer to limit the number of
iterations, to NULL or 0 (default) for no limits}

\item{warnings}{Logical value, to display the warnings and error raised by
the underlying functions}
}
\value{
The function returns an S3 object of type \code{BRIL.MedianRec},
containing the following values:
\item{\code{median}}{Coordinate of the recursive median}
\item{\code{max}}{Coordinate of the sample with the highest depth (or the
center of the first iteration in the case of convex body minimizers)}
\item{\code{iterations}}{List containing the indices from the samples of \code{data}
selected at each iteration}
}
\description{
Recursive estimate of central location based on depth measures (from the
packages
\href{https://CRAN.R-project.org/package=depth}{\code{depth}} and
\href{https://CRAN.R-project.org/package=ddalpha}{\code{ddalpha}})
or convex body minimizers (package
\href{https://CRAN.R-project.org/package=MASS}{\code{MASS}}).
}
\examples{

# Illustrative data
XY <- rbind(
  mvtnorm::rmvnorm(300, c(0, 0), diag(2)),
  mvtnorm::rmvnorm(100, c(15, 20), diag(2) * 3 - 1),
  mvtnorm::rmvnorm(150, c(-10, 15), diag(2) * 2 - 0.5),
  mvtnorm::rmvnorm(200, c(5, 5), diag(2) * 200)
)

# Compute the recursive median
res <- median_rec(XY)

print(res)
plot(res)

}
\seealso{
\code{\link[=plot.BRIL.MedianRec]{plot.BRIL.MedianRec()}}, \code{\link[=print.BRIL.MedianRec]{print.BRIL.MedianRec()}}, \code{\link[=median_mv]{median_mv()}}, \code{\link[=depth_values]{depth_values()}}, \code{\link[=bril]{bril()}}
}
