% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/median_rec.R
\name{plot.BRIL.MedianRec}
\alias{plot.BRIL.MedianRec}
\title{Plot method for \code{BRIL.MedianRec} objects}
\usage{
\method{plot}{BRIL.MedianRec}(x, nbIterations = 5, showMedian = FALSE, showMax = FALSE, ...)
}
\arguments{
\item{x}{An object of class \code{BRIL.MedianRec} (see \code{\link[=median_rec]{median_rec()}})}

\item{nbIterations}{Number of iterations to display, or 0 to show all of them
(default: 5)}

\item{showMedian}{Logical value, to show the final recursive median
(indicated by a "+")}

\item{showMax}{Logical value, to show the overall deepest point, or the
center of the first MCD/MVE iteration (indicated by a "x")}

\item{...}{Other arguments passed to or from other methods}
}
\description{
Plot method for \code{BRIL.MedianRec} objects
}
\examples{

# Illustrative data
XY <- rbind(
  mvtnorm::rmvnorm(300, c(0, 0), diag(2)),
  mvtnorm::rmvnorm(100, c(15, 20), diag(2) * 3 - 1),
  mvtnorm::rmvnorm(150, c(-10, 15), diag(2) * 2 - 0.5),
  mvtnorm::rmvnorm(200, c(5, 5), diag(2) * 200)
)

# Process the data
res <- median_rec(XY)

# Default plot
plot(res)

# Adjust axis
plot(res, asp = 1, xlim = c(-20, 20), ylim = c(-20, 20))

# Change other graphical options
plot(res, showMedian = TRUE, pch = 16, main = "Recursive Median")

}
\seealso{
\code{\link[=median_rec]{median_rec()}}, \code{\link[=median_mv]{median_mv()}}, \code{\link[=bril]{bril()}}
}
