% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bril.R
\name{plot.BRIL}
\alias{plot.BRIL}
\title{Plot method for \code{BRIL} objects}
\usage{
\method{plot}{BRIL}(
  x,
  contents = "plot",
  showClusters = TRUE,
  showMode = TRUE,
  col,
  colMode,
  colClusters = NULL,
  iterationsIndices = NULL,
  iterationsOptions = NULL,
  ...
)
}
\arguments{
\item{x}{An object of class \code{BRIL} (see \code{\link[=bril]{bril()}})}

\item{contents}{Contents to be displayed, options are "scatterplot",
or "iterations" (only one option possible)}

\item{showClusters}{Logical value. Used with contents="scatterplot", to show
or not the different clusters}

\item{showMode}{Logical value. Used for contents="clusters", to show or not
the main mode to show or not the different clusters}

\item{col}{default color of samples}

\item{colMode}{Color of the mode when contents="scatterplot" (if showMode
is TRUE)}

\item{colClusters}{List (or array) of colors (length must be at least equal
to the number of clusters identified by the function \code{\link[=bril]{bril()}}, i.e.
`x$nbClusters)}

\item{iterationsIndices}{Numerical value or array of numerical values, used
with contents="iterations", providing the indices of the iterations to be
ploted. If more than one iteration is requested, an interactive menu in the
console will be used for the selection. 0 or NULL (default) will include
all the iterations. Values that are negative or superior to the number of
iterations performed by the execution \code{\link[=bril]{bril()}} will be ignored}

\item{iterationsOptions}{List of additional parameters to be passed to the
\code{plot.BRIL.Filtering} function when contents="iterations" is selected (see
\link{plot.BRIL.Filtering} for details). \emph{Example: iterationsOptions =
list(Xlab=NA, Ylab=NA, contents=c("p.values", "scatterplot"))}}

\item{...}{Other arguments passed to or from other methods (such as pch for
the symbols, main and sub for title and subtitle, xlab, xmin, ...)}
}
\description{
Plot method for \code{BRIL} objects
}
\seealso{
\code{\link[=bril]{bril()}}, \link{print.BRIL}, \code{\link[=filter_outliers]{filter_outliers()}}, \code{\link[=print.BRIL.Filtering]{print.BRIL.Filtering()}}
}
