% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bril.R
\name{print.BRIL}
\alias{print.BRIL}
\title{Print method for \code{BRIL} objects}
\usage{
\method{print}{BRIL}(x, maxDisplayed = NULL, ...)
}
\arguments{
\item{x}{An object of class \code{BRIL} (see \code{\link[=bril]{bril()}})}

\item{maxDisplayed}{Number of elements to display in the output. Set to NULL
(or 0) to show all values.}

\item{...}{Other arguments passed to or from other methods}
}
\description{
Print method for \code{BRIL} objects
}
\seealso{
\code{\link[=plot.BRIL]{plot.BRIL()}}, \code{\link[=bril]{bril()}}
}
