% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_outliers.R
\name{test_multinormality}
\alias{test_multinormality}
\title{Test the multivariate normality of a distribution}
\usage{
test_multinormality(
  values,
  test = "Mardia",
  threshold = 0.05,
  data,
  center,
  warnings = FALSE
)
}
\arguments{
\item{values}{Unidimentional array of numerical values (distances)}

\item{test}{Statistical test used (for now the only option "DIP")}

\item{threshold}{Threshold of significance for the statistical test
(default: 0.05)}

\item{data}{Matrix of numerical values containing the observations (one per
row, with two columns for X and Y coordinates)}

\item{center}{Center of the observations from \code{data}}

\item{warnings}{Logical value, to display the warnings and errors caught}
}
\value{
p-value of the test (lower values suggest non normality)
}
\description{
Test the multivariate normality of a distribution
}
\details{
Parameter \code{data} is only required for the tests "Mardia",
"Skewness", "Kurtosis" and "Chisq", while parameter \code{center} is only
required for "Chisq"
}
