% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_outliers.R
\name{test_unimodality}
\alias{test_unimodality}
\title{Test the unimodality of a distribution}
\usage{
test_unimodality(values, test = "DIP", warnings = FALSE)
}
\arguments{
\item{values}{Unidimensional array of numerical values (distances)}

\item{test}{Statistical test used (for now the only option is "DIP")}

\item{warnings}{Logical value, to display the warnings and errors caught}
}
\value{
p-value of the test (lower values suggest multimodality)
}
\description{
Test the unimodality of a distribution
}
\seealso{
\code{\link[=filter_outliers]{filter_outliers()}}
}
